Fixing ai slop
