+++ gaze-security/Cargo.toml
@@ -17,6 +17,3 @@ libc = "0.2.189"
 sha2 = "0.11"
 tss-esapi = "7.7.0"
 zeroize = "1.8"
-
-[dev-dependencies]
-tempfile = "3"
