+++ gaze/Cargo.toml
@@ -19,7 +19,6 @@ default = []
 openvino = ["gaze-core/openvino", "ort/openvino"]
 
 [dependencies]
-gaze-security = { path = "../gaze-security" }
 aes-gcm = "0.11.1"
 anyhow = "1.0.104"
 gaze-core = { path = "../gaze-core" }
@@ -48,6 +47,7 @@ tracing = "0.1.44"
 tracing-subscriber = { version = "0.3.23", features = ["env-filter"] }
 futures = "0.3.34"
 zbus_polkit = "5.0.0"
+tss-esapi = "7.7.0"
 
 [[bin]]
 name = "gazed"
